refactor: split ActualGame component in two pages improved game joining process
