+++ app/game/_components/SkeletonGame.tsx
@@ -28,9 +28,9 @@ export default function SkeletonGame({
 
   return (
     <section className="mx-auto flex max-w-screen-board flex-col items-center justify-center">
-      {exceptionFromBackendChessService ? (
+      {exceptionFromBackendChessService && (
         <p>{exceptionFromBackendChessService.message}</p>
-      ) : null}
+      )}
       <UserInfo isLoading />
       <div className="relative w-full">
         <div
